Refactor neo/session
